refactor: Feedback from code-review
