+++ examples/external-data/package.json
@@ -17,12 +17,14 @@
 		"build:compile": "fluid-build . --task compile",
 		"build:esnext": "tsc --project ./tsconfig.json",
 		"build:test": "npm run build:test:esm",
-		"build:test:esm": "tsc --project ./tests/tsconfig.json",
+		"build:test:esm": "npm run build:test:mocha && npm run build:test:jest",
+		"build:test:jest": "tsc --project ./test/tsconfig.json",
+		"build:test:mocha": "tsc --project ./src/test/tsconfig.json",
 		"check:biome": "biome check .",
 		"check:format": "npm run check:biome",
 		"clean": "rimraf --glob dist lib \"**/*.tsbuildinfo\" \"**/*.build.log\" nyc",
-		"eslint": "eslint --quiet --format stylish src tests",
-		"eslint:fix": "eslint --quiet --format stylish src tests --fix",
+		"eslint": "eslint --quiet --format stylish src test",
+		"eslint:fix": "eslint --quiet --format stylish src test --fix",
 		"format": "npm run format:biome",
 		"format:biome": "biome check . --write",
 		"good-fences": "gf",
@@ -36,8 +38,8 @@
 		"start:app:test": "start-server-and-test start:services \"5236|5237\" start:client:test",
 		"start:client": "webpack serve",
 		"start:client:test": "webpack serve --config webpack.test.cjs",
-		"start:customer-service": "cross-env CUSTOMER_SERVICE_PORT=5237 node dist/src/mock-customer-service/start.js",
-		"start:external-data-service": "cross-env EXTERNAL_DATA_SERVICE_PORT=5236 node dist/src/mock-external-data-service/start.js",
+		"start:customer-service": "cross-env CUSTOMER_SERVICE_PORT=5237 node dist/mock-customer-service/start.js",
+		"start:external-data-service": "cross-env EXTERNAL_DATA_SERVICE_PORT=5236 node dist/mock-external-data-service/start.js",
 		"start:services": "start-server-and-test start:external-data-service 5236 start:customer-service",
 		"start:test": "start-server-and-test start:tinylicious 7070 start:app:test",
 		"start:tinylicious": "tinylicious",
@@ -113,6 +115,7 @@
 		"jest-puppeteer": "^10.1.3",
 		"jiti": "^2.6.1",
 		"mocha": "^11.7.5",
+		"mocha-multi-reporters": "^1.5.1",
 		"process": "^0.11.10",
 		"puppeteer": "^23.6.0",
 		"rimraf": "^6.1.3",
