+++ examples/external-data/src/test/externalDataService.test.ts
@@ -14,9 +14,9 @@ import {
 	ExternalDataSource,
 	type MockWebhook,
 	initializeExternalDataService,
-} from "../src/mock-external-data-service/index.js";
-import { externalDataServicePort } from "../src/mock-external-data-service-interface/index.js";
-import { type ITaskData, assertValidTaskData } from "../src/model-interface/index.js";
+} from "../mock-external-data-service/index.js";
+import { externalDataServicePort } from "../mock-external-data-service-interface/index.js";
+import { type ITaskData, assertValidTaskData } from "../model-interface/index.js";
 
 import { closeServer, delay } from "./utilities.js";
