+++ examples/external-data/src/test/utilities.ts
