+++ examples/external-data/test/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "../../../common/build/build-common/tsconfig.test.node16.json",
+	"compilerOptions": {
+		"rootDir": "./",
+		"outDir": "../lib/test-app",
+		"types": ["jest", "puppeteer", "jest-environment-puppeteer", "expect-puppeteer", "react"],
+	},
+	"include": ["./**/*"],
+	"references": [
+		{
+			"path": "..",
+		},
+	],
+	"lib": ["dom"],
+}
