--- examples/external-data/tests/tsconfig.json
@@ -1,19 +1,0 @@
-{
-	"extends": "../../../common/build/build-common/tsconfig.test.node16.json",
-	"compilerOptions": {
-		"rootDir": "./",
-		"outDir": "../lib/test",
-		"types": ["mocha", "node"],
-	},
-	"include": [
-		"./customerService.test.ts",
-		"./externalDataService.test.ts",
-		"./taskData.test.ts",
-		"./utilities.ts",
-	],
-	"references": [
-		{
-			"path": "..",
-		},
-	],
-}
