+++ examples/external-data/tsconfig.json
@@ -1,11 +1,12 @@
 {
 	"extends": "../../common/build/build-common/tsconfig.node16.json",
 	"compilerOptions": {
+		"rootDir": "./src",
 		"outDir": "./lib",
-		"types": ["jest", "puppeteer", "jest-environment-puppeteer", "expect-puppeteer", "react"],
 		"noUncheckedIndexedAccess": false,
 		"exactOptionalPropertyTypes": false,
 	},
-	"include": ["src/**/*", "tests/index.tsx", "tests/taskList.test.ts"],
+	"include": ["src/**/*"],
+	"exclude": ["src/test/**"],
 	"lib": ["dom"],
 }
